+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added package documentation ([issue #5](https://github.com/ocbe-uio/permChacko/issues/5)).
 * Standardized dataset names
 * Created print methods for `permChacko()` and `reduceVector()` ([issue #7](https://github.com/ocbe-uio/permChacko/issues/7)).
+* Created summary method for `reduceVector()` ([issue #8](https://github.com/ocbe-uio/permChacko/issues/8)).
 * Added vignette ([issue #2](https://github.com/ocbe-uio/permChacko/issues/2)).
 
 # permChacko 0.1.0